Carlos araujo continuando alterações em exchanges

+++ back-end/src/controllers/exchanges.controller.ts
@@ -2,8 +2,9 @@ import { Request, Response } from 'express';
 import exchangesService from '../services/exchanges.service';
 import statusCodes from '../statusCodes';
 
-const getAllExchanges = async (req: Request, res: Response) => {
-  const exchanges = await exchangesService.getAllExchanges();
+const getAllExchangesByReader = async (req: Request, res: Response) => {
+  const { id } = req.body.user;
+  const exchanges = await exchangesService.getAllExchangesByReader(id);
   res.status(statusCodes.OK).json(exchanges);
 };
 
@@ -23,12 +24,33 @@ const createExchange = async (req: Request, res: Response) => {
   return res.status(statusCodes.CREATED).json(newExchange);
 }
 
+const confirmExchange = async (req: Request, res: Response) => {
+  const { id } = req.params;
+  const { id: readerId } = req.body.user;
+  const result = await exchangesService.getExchangeById(Number(id));
+  if (!result) {
+    return res.status(statusCodes.NOT_FOUND).json({ message: 'Exchange not found'});
+  }
+  if (result.receiverId != readerId) {
+    return res.status(statusCodes.UNAUTHORIZED).json({ message: 'Reader cannot confirm exchange' });
+  }
+  const updatedQty = await exchangesService.confirmExchange(Number(id));
+  if (updatedQty) {
+    const result = await exchangesService.getExchangeById(Number(id));
+    return res.status(statusCodes.OK).json(result);
+  }
+  return res.status(statusCodes.ERROR).json({ message: 'Error'});
+}
+
 const deleteExchange = async ( req: Request, res: Response) => {
   const { id } = req.params;
   const { id: readerId } = req.body.user;
   const result = await exchangesService.getExchangeById(Number(id));
   if (!result) {
     return res.status(statusCodes.NOT_FOUND).json({ message: 'Exchange not found'});
+  }
+  if (result.receiverId != readerId) {
+    return res.status(statusCodes.UNAUTHORIZED).json({ message: 'Reader cannot delete exchange' });
   }
   const exchange = await exchangesService.deleteExchange(Number(id));
   if (exchange) {
@@ -38,4 +60,8 @@ const deleteExchange = async ( req: Request, res: Response) => {
   
 };
 
-export default { getAllExchanges, createExchange, deleteExchange, getExchangeById };
+export default { getAllExchangesByReader, 
+  createExchange, 
+  deleteExchange, 
+  getExchangeById,
+  confirmExchange };
